Development of Shiny
Development of UI for simulation

Current issue: the counters for the functions are not inherited to the
environments in which the functions are excecuted, this breaks a lot of
things, all checks regarding years and making the output files

Proposed solution: make the counter arguments in the functions and
create the variables within the functions from the arguments. That
should solve the issue. Its only three things, p,y and year. They are
called from a multitude of functions and are therfore bothersome to do.