debugging
There is something fishy about the correlation of blup to breeding
values

Something is wrong somewhere. Could be the model, inbreeding, setting
thing for dmu (inbreeding structure for the pedigree), could be a code
bug somewhere. variance is dropping while the phenotypic variance stays
the same, this means that the heritability is dropping proposterously
fast.

Need to figure this out!