Working new correlation matrix
This version has three traits for body weight. Now shifting to RR model
for BW
Still an issue with the extreme heritability of BW in that model...
Don't know if i'll put an heritability of 0.82 in the model with the
negative correlations to litter size. It will completely kill it.