Improvements on controls
Fixed functions so control on Simulation function actually works as
intended. Now the constants specified in Simulation function actually
are used within it. This means that changing values in the definition
file is now redundant, other than the genetic constants which are are
not intended to be messed with and are not in the argument list in the
Simulation function