Version 20.0
Fix of labor cost model and addition of a cell that keeps track of only
labor costs

ALSO fixed pretty major bug in feed use of breeders, it did not
calculate the feed used for lactation. Not sure how much of an impact
that has on litter size economic values. Need to rerun all of those. It
should not affect the body weight since it is only based on numbers of
kits in litter